# trip-builder/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trip Builder API",
        "description": "Simple trip builder api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "api server"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs user into system",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/x-www-form-urlencoded": {},
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logs out current logged in user session",
                "operationId": "logout",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/cities": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get list of cities",
                "description": "Returns list of cities",
                "operationId": "getCitiesList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Store city",
                "description": "Create new city",
                "operationId": "postCity",
                "requestBody": {
                    "description": "City object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/City"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/cities/{city}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Get city information",
                "description": "Returns city data",
                "operationId": "getCityById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cities"
                ],
                "summary": "Update city",
                "description": "update an city",
                "operationId": "putCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "City object",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/City"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cities"
                ],
                "summary": "Delete city",
                "description": "delete an city",
                "operationId": "deleteCity",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "City": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "code"
                        ],
                        "properties": {
                            "code": {},
                            "name": {},
                            "region_id": {}
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}